Refactor UI package structure
The current UI package structure is not intuitive and extensible.
OOP principles are violated as the the enclosing component is responsible
for loading the subcomponents of their child's UI. This makes it confusing
to the programmer. For example, the MainWindow class is the responsibility
of loading each tab into the TabPane.

The java files are also all thrown into the same package in the ui package
which makes it difficult to navigate and organise the ui components according
to their hierachy.

This commit refactors the UI package structure to be more intuitive and
extensible. The following changes were made:
1. Renamed ResultDisplay to CommandLog
1. Abstracted out TabsPane into the TabsHolder class
1. Moved all UI parts into their own subpackages
1. Moved UI test files to the same package as the UI files they test

With this refactor, the UI package structure is now more intuitive and extensible.
