+++ src/main/java/mycelium/mycelium/ui/Ui.java
@@ -1,4 +1,4 @@
-package mycelium.mycelium.ui.common;
+package mycelium.mycelium.ui;
 
 import javafx.stage.Stage;
