+++ src/main/java/mycelium/mycelium/ui/UiPart.java
@@ -1,4 +1,4 @@
-package mycelium.mycelium.ui.common;
+package mycelium.mycelium.ui;
 
 import static java.util.Objects.requireNonNull;
