+++ src/main/java/mycelium/mycelium/ui/statusbarfooter/StatusBarFooter.java
@@ -1,4 +1,4 @@
-package mycelium.mycelium.ui;
+package mycelium.mycelium.ui.statusbarfooter;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -6,7 +6,7 @@ import java.nio.file.Paths;
 import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.Region;
-import mycelium.mycelium.ui.common.UiPart;
+import mycelium.mycelium.ui.UiPart;
 
 /**
  * A ui for the status bar that is displayed at the footer of the application.
