+++ src/main/resources/view/CommandLog.fxml
