+++ src/test/java/mycelium/mycelium/ui/UiPartTest.java
@@ -13,7 +13,6 @@ import org.junit.jupiter.api.io.TempDir;
 
 import javafx.fxml.FXML;
 import mycelium.mycelium.MainApp;
-import mycelium.mycelium.ui.common.UiPart;
 
 public class UiPartTest {
