+++ src/test/java/mycelium/mycelium/ui/resultoutput/ClientListCardTest.java
@@ -1,4 +1,4 @@
-package mycelium.mycelium.ui.common.client;
+package mycelium.mycelium.ui.resultoutput;
 
 import static mycelium.mycelium.ui.testutil.GuiTestAssert.assertCardDisplaysClient;
 import static org.junit.jupiter.api.Assertions.assertEquals;
